requirements.txt
flask==3.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
nltk==3.8.1
plotly==5.18.0
transformers==4.36.2
python-dotenv==1.0.0
flask-cors==4.0.0
